fix(control): add safety check to ngOnDestroy
ref #53

+++ projects/ngx-mapbox-gl/src/lib/control/control.component.ts
@@ -56,6 +56,8 @@ export class ControlComponent implements OnDestroy, AfterContentInit {
   }
 
   ngOnDestroy() {
-    this.MapService.removeControl(this.control);
+    if (this.control) {
+      this.MapService.removeControl(this.control);
+    }
   }
 }
